requirements.txt
google-genai
flask
gunicorn
requests
python-dotenv
